+++ apps/dashboard/package.json
@@ -26,7 +26,7 @@
     "@midday/utils": "workspace:*",
     "@novu/headless": "^0.24.0",
     "@openstatus/react": "^0.0.3",
-    "@tanstack/react-table": "^8.13.2",
+    "@tanstack/react-table": "^8.14.0",
     "@todesktop/client-active-win": "^0.15.0",
     "@todesktop/client-core": "^1.2.5",
     "@todesktop/runtime": "^1.6.1",
@@ -49,7 +49,7 @@
     "loops": "0.3.0",
     "lottie-react": "^2.4.0",
     "ms": "^2.1.3",
-    "next": "14.1.3",
+    "next": "14.1.4",
     "next-international": "1.2.4",
     "next-safe-action": "6.2.0",
     "next-themes": "^0.3.0",
