+++ apps/website/package.json
@@ -20,7 +20,7 @@
     "@vercel/analytics": "^1.2.2",
     "date-fns": "^3.6.0",
     "geist": "^1.2.2",
-    "next": "14.1.3",
+    "next": "14.1.4",
     "next-international": "1.2.4",
     "next-themes": "^0.3.0",
     "react": "18.2.0",
@@ -30,8 +30,8 @@
   },
   "devDependencies": {
     "@midday/tsconfig": "workspace:*",
-    "@types/node": "^20.11.28",
-    "@types/react": "^18.2.66",
+    "@types/node": "^20.11.30",
+    "@types/react": "^18.2.67",
     "@types/react-dom": "^18.2.22"
   }
 }
