+++ apps/website/src/app/[locale]/oss-friends/page.tsx
@@ -1,14 +1,9 @@
-import { getStaticParams } from "@/locales/server";
 import type { Metadata } from "next";
 import { setStaticParamsLocale } from "next-international/server";
 
 export const metadata: Metadata = {
   title: "OSS Friends | Midday",
 };
-
-export function generateStaticParams() {
-  return getStaticParams();
-}
 
 export default async function Page({
   params: { locale },
