+++ apps/website/src/app/[locale]/page.tsx
@@ -1,10 +1,5 @@
 import { StartPage } from "@/components/startpage";
-import { getStaticParams } from "@/locales/server";
 import { setStaticParamsLocale } from "next-international/server";
-
-export function generateStaticParams() {
-  return getStaticParams();
-}
 
 export default function Page({
   params: { locale },
