+++ apps/website/src/app/[locale]/support/page.tsx
@@ -1,14 +1,9 @@
-import { getStaticParams } from "@/locales/server";
 import type { Metadata } from "next";
 import { setStaticParamsLocale } from "next-international/server";
 
 export const metadata: Metadata = {
   title: "Support | Midday",
 };
-
-export function generateStaticParams() {
-  return getStaticParams();
-}
 
 export default function Page({
   params: { locale },
