+++ apps/website/src/app/[locale]/talk-to-us/page.tsx
@@ -1,15 +1,10 @@
 import { CalEmbed } from "@/components/cal-embed";
-import { getStaticParams } from "@/locales/server";
 import type { Metadata } from "next";
 import { setStaticParamsLocale } from "next-international/server";
 
 export const metadata: Metadata = {
   title: "Talk to us | Midday",
 };
-
-export function generateStaticParams() {
-  return getStaticParams();
-}
 
 export default function Page({
   params: { locale },
