+++ apps/website/src/app/[locale]/terms/page.tsx
@@ -1,14 +1,9 @@
-import { getStaticParams } from "@/locales/server";
 import type { Metadata } from "next";
 import { setStaticParamsLocale } from "next-international/server";
 
 export const metadata: Metadata = {
   title: "Terms and Conditions | Midday",
 };
-
-export function generateStaticParams() {
-  return getStaticParams();
-}
 
 export default function Page({
   params: { locale },
