+++ apps/website/src/middleware.ts
@@ -12,7 +12,5 @@ export function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: [
-    "/((?!.*\\.|api|_next/static|_next/image|favicon.ico|robots.txt|auth\\/).*)",
-  ],
+  matcher: ["/((?!api|static|.*\\..*|_next|favicon.ico|robots.txt).*)"],
 };
